bolt11: move to common/ and sign via callback.
JSON stuff is moved to lightningd/invoice.c.
