+++ common/Makefile
@@ -1,6 +1,7 @@
 COMMON_SRC :=					\
 	common/bech32.c				\
 	common/bip32.c				\
+	common/bolt11.c				\
 	common/channel_config.c			\
 	common/close_tx.c			\
 	common/configdir.c			\
